Retain the upstream typed record, CLIPS recovery and alignment repairs alongside
local unit metadata, tactical fields, tooling tests and generated consumer headers.
Resolve the vision-radius function using typed slots and full metadata reads.
Refresh only nine reviewed canonical hashes, including eight stale upstream
hashes left by the packed-pointer alignment repair. Fix clean-checkout doc links.

Validation: 4157-function/140-TU split audit, generator freshness, 45 tooling
tests, and the standalone CLASH.DAT checks pass. Fresh dual-compiler and runtime
validation follows the remaining naming integration; no campaign promotion.

+++ src/clips/0049B530_0049CD60_clips_004.c
@@ -1248,7 +1248,7 @@ int  Rules_WhileFunction(double context)
 
   ++g_ClipsCurrentEvaluationDepth;
   Rules_RtnUnknown(1, theResult, context);
-  while ( theResult[2] != g_ClipsFalseSymbol || theResult[1] != 2 )
+  while ( (int)theResult[2] != g_ClipsFalseSymbol || theResult[1] != 2 )
   {
     if ( g_ClipsHaltExecution == 1 )
       break;
